+++ tsup.config.ts
@@ -7,6 +7,7 @@ export default defineConfig({
     index: './src/index.ts',
     time: './src/time/index.ts',
     coordinates: './src/coordinates/index.ts',
+    frames: './src/frames/index.ts',
     constants: './src/constants/index.ts',
     rinex: './src/rinex/index.ts',
     rtcm3: './src/rtcm3/index.ts',